Lwdev 9178 fix address history updating

+++ src/Lykke.Service.Qtum.Api.Core/Domain/InsightApi/IVin.cs
@@ -16,7 +16,7 @@
 
         long ValueSat { get; set; }
 
-        double Value { get; set; }
+        decimal Value { get; set; }
 
         object DoubleSpentTxId { get; set; }
     }
